requirements.txt
telethon
PySocks